chore: rename components and adjust design
