--- src/components/Stepper.tsx
@@ -1,21 +1,0 @@
-type Props = {
-  width?: string
-};
-
-export default function Stepper(
-  {
-    width,
-  }: Props,
-) {
-  return (
-    <div
-      style={{ width: `${width}px` }}
-      className="stepper d-flex position-relative justify-content-between mb-4"
-    >
-      <span className="dot active" />
-      <span className="dot" />
-      <span className="dot" />
-      <span className="dot" />
-    </div>
-  );
-}
