+++ src/styles/base.scss
@@ -85,32 +85,4 @@
       transform: rotate(-45deg);
     }
   }
-
-  // stepper
-
-  .stepper {
-    &::before {
-      position: absolute;
-      top: 50%;
-      left: 5px;
-      z-index: 0;
-      width: calc(100% - 10px);
-      height: 1px;
-      content: "";
-      background-color: #fff;
-    }
-    > .dot {
-      position: relative;
-      z-index: 1;
-      width: 1rem;
-      height: 1rem;
-      background: #161723;
-      border: 1px solid #fff;
-      border-radius: 50%;
-      &.active {
-        background-color: #fff;
-        border-color: #fff;
-      }
-    }
-  }
 }
